backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
SQLAlchemy
python-pptx
openai
